Dev_19
FireFox has this weird bug where textarea text does not fill the textarea.

It also has a very weird bug where the above bug only happens on the page's first textarea, as shown in our code. The 2nd and 3rd textareas fill correctly.

The workaround is to make a dummy textarea, hide it with Angular, then continue on on our merry way.